chore: organized seedfarmer.models; added ProjectSpec schema for seedfarmer.yaml
