+++ seedfarmer/models/manifests/_deployment_manifest.py
@@ -13,135 +13,12 @@
 #    limitations under the License.
 
 import os
-from enum import Enum
 from typing import Any, Dict, List, Optional, Tuple, Union, cast
 
 from pydantic import PrivateAttr
 
-from seedfarmer.models._base import CamelModel
-from seedfarmer.utils import upper_snake_case
-
-
-class BuildType(Enum):
-    BUILD_GENERAL1_SMALL = "BUILD_GENERAL1_SMALL"
-    BUILD_GENERAL1_MEDIUM = "BUILD_GENERAL1_MEDIUM"
-    BUILD_GENERAL1_LARGE = "BUILD_GENERAL1_LARGE"
-    BUILD_GENERAL1_2XLARGE = "BUILD_GENERAL1_2XLARGE"
-
-
-class BuildPhase(CamelModel):
-    """
-    BuildPhase
-    This is a list of strings that are passed to CodeSeeder to be executed
-    in their respective phases as commands
-    """
-
-    commands: List[str] = []
-
-
-class BuildPhases(CamelModel):
-    """
-    BuildPhases
-    This object has the individual commands for each of the define build phases:
-    install, pre_build,  build, post_build
-    """
-
-    install: BuildPhase = BuildPhase()
-    pre_build: BuildPhase = BuildPhase()
-    build: BuildPhase = BuildPhase()
-    post_build: BuildPhase = BuildPhase()
-
-
-class ExecutionType(CamelModel):
-    """
-    ExecutionType
-    This an object that contains the Build Phases object for the destroy or deploy
-    object of the DeploySpec
-    """
-
-    phases: BuildPhases = BuildPhases()
-
-
-class DeploySpec(CamelModel):
-    """
-    DeploySpec
-    This represents the commands passed to CodeSeeder that will be executed
-    on behalf of the module to be built.
-    The deploy and the destroy objects each have an ExecutionType object.
-    """
-
-    deploy: Optional[ExecutionType] = None
-    destroy: Optional[ExecutionType] = None
-    build_type: Optional[str] = None
-    publish_generic_env_variables: Optional[bool] = False
-
-    def __init__(self, **data: Any) -> None:
-        super().__init__(**data)
-        if data.get("build_type"):
-            chk = str(data["build_type"]).upper()
-            self.build_type = chk if chk in BuildType.__members__.keys() else BuildType.BUILD_GENERAL1_SMALL.value
-        else:
-            self.build_type = BuildType.BUILD_GENERAL1_SMALL.value
-
-
-class ModuleRef(CamelModel):
-    name: str
-    group: str
-    key: Optional[str] = None
-
-
-class ValueRef(CamelModel):
-    module_metadata: Optional[ModuleRef] = None
-    env_variable: Optional[str] = None
-    parameter_store: Optional[str] = None
-    secrets_manager: Optional[str] = None
-    parameter_value: Optional[str] = None
-
-
-class ValueFromRef(CamelModel):
-    value_from: Optional[ValueRef] = None
-
-
-class ModuleParameter(ValueFromRef):
-    _upper_snake_case: str = PrivateAttr()
-
-    name: str
-    value: Optional[Any] = None
-
-    def __init__(self, **data: Any) -> None:
-        super().__init__(**data)
-        self._upper_snake_case = upper_snake_case(self.name)
-
-    @property
-    def upper_snake_case(self) -> str:
-        return self._upper_snake_case
-
-
-class ModuleManifest(CamelModel):
-    """
-    ModuleManifest
-    This is a module in a group of the deployment and consists of a name,
-    the path of the module manifest, any ModuleParameters for deployment, and
-    the DeploySpec.
-    """
-
-    name: str
-    path: str
-    bundle_md5: Optional[str]
-    manifest_md5: Optional[str]
-    deployspec_md5: Optional[str]
-    parameters: List[ModuleParameter] = []
-    deploy_spec: Optional[DeploySpec] = None
-    target_account: Optional[str] = None
-    target_region: Optional[str] = None
-    codebuild_image: Optional[str] = None
-    _target_account_id: Optional[str] = PrivateAttr(default=None)
-
-    def set_target_account_id(self, account_id: str) -> None:
-        self._target_account_id = account_id
-
-    def get_target_account_id(self) -> Optional[str]:
-        return self._target_account_id
+from seedfarmer.models._base import CamelModel, ValueFromRef
+from seedfarmer.models.manifests._module_manifest import ModuleManifest
 
 
 class ModulesManifest(CamelModel):
